+++ CHANGES.md
@@ -1,5 +1,7 @@
 # 1.9.1
 - fix: wrong date localization at beginning of year
+- fix: achievement for month goal streak: fix calculation of streak
+- fix: achievement for month goal streak: current month should be added to streak if all goals are reached even if month is not over yet
 
 # 1.9.0
 - add: localize dates, numbers and maps